Update20/11/2019

+++ src/L4/Hotel.java
@@ -16,7 +16,7 @@ public class Hotel extends L4 {
 
     public Hotel(){}
 
-    Hotel(String name, String location, double rating, String[] RoomDetails, int[] noOfRooms){
+    Hotel(String name, String location, int rating, String[] RoomDetails, int[] noOfRooms){
         this.name = name;
         this.location = location;
         this.rating = rating;
